requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via supervision
distlib==0.3.8
    # via virtualenv
filelock==3.13.1
    # via
    #   torch
    #   virtualenv
fonttools==4.49.0
    # via matplotlib
fsspec==2024.2.0
    # via torch
identify==2.5.35
    # via pre-commit
idna==3.6
    # via requests
jinja2==3.1.3
    # via torch
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via
    #   seaborn
    #   supervision
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
nodeenv==1.8.0
    # via pre-commit
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   scipy
    #   seaborn
    #   supervision
    #   torchvision
opencv-python==4.9.0.80
    # via
    #   -r requirements.in
    #   ultralytics
opencv-python-headless==4.9.0.80
    # via supervision
packaging==23.2
    # via
    #   build
    #   matplotlib
pandas==2.2.0
    # via
    #   -r requirements.in
    #   seaborn
    #   ultralytics
pillow==10.2.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
pip-tools==7.4.0
    # via -r requirements.in
platformdirs==4.2.0
    # via virtualenv
pre-commit==3.6.2
    # via -r requirements.in
psutil==5.9.8
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pyparsing==3.1.1
    # via matplotlib
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   pre-commit
    #   supervision
    #   ultralytics
requests==2.31.0
    # via
    #   torchvision
    #   ultralytics
scipy==1.12.0
    # via
    #   supervision
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
six==1.16.0
    # via python-dateutil
supervision==0.18.0
    # via -r requirements.in
sympy==1.12
    # via torch
thop==0.1.1.post2209072238
    # via ultralytics
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
torch==2.2.0
    # via
    #   thop
    #   torchvision
    #   ultralytics
torchvision==0.17.0
    # via ultralytics
tqdm==4.66.2
    # via ultralytics
typing-extensions==4.9.0
    # via torch
tzdata==2024.1
    # via pandas
ultralytics==8.1.16
    # via -r requirements.in
urllib3==2.2.1
    # via requests
virtualenv==20.25.1
    # via pre-commit
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
